Clean up old controls
Remove unused controls from Authoring/Studio/Controls and some
ancillary files elsewhere.

Task-number: QT3DS-1828

+++ src/Authoring/Studio/Palettes/Inspector/InspectorControlModel.cpp
@@ -307,7 +307,7 @@ InspectorControlBase* InspectorControlModel::createMaterialItem(Qt3DSDMInspectab
         break;
     }
 
-    for (size_t matIdx = 0, end = m_materials.size(); matIdx < end; ++matIdx) {
+    for (int matIdx = 0, end = int(m_materials.size()); matIdx < end; ++matIdx) {
         if (m_materials[matIdx].m_relativePath == sourcePath)
             item->m_value = values[matIdx + 2]; // +2 for standard and referenced materials
     }
